+++ docs/userguide/docs/changes.md
@@ -9,7 +9,7 @@ Not yet released
 
 Minor enhancements
 * Reworked [distribution support API](basics.md#randomness--distributions) for better API experience to enable controlled randomization in process models
-* Better API support for [depletable resources](resource.md#depletable-resources) 
+* [#38](https://github.com/holgerbrandl/kalasim/issues/38) Better API support for [depletable resources](resource.md#depletable-resources) 
 
 ## v0.7